[XLA:GPU] Switch from specifying maxntid to reqntid.
maxntid specifies the max number of threads in a block, whereas reqntid
says that we will use *exactly* this many threads in a block.

This doesn't have any effect on the benchmarks I ran, but we might as
well do it in case it helps ptxas make a better decision at some point
on some GPU.  At least it will prevent the next person to come along
from doing this same investigation I just did.  :)

PiperOrigin-RevId: 177851116