[StreamExecutor] Change "variance" to "inv_var" in BatchNormalizationBackward.
This parameter is not the variance of the data, but rather is
1/(sqrt(variance + epsilon).  Neglecting epsilon, this is the inverse
standard deviation.

"inv_stddev" might be a better name, but "inv_var" is certainly better
than plain "variance", and it matches nvidia's name for this parameter,
which I think may override the desire for a more precise name.

No functional change.

PiperOrigin-RevId: 179352839